fix(gallery): handle large media files safely in thumbnail generation

+++ src/native/include/bridge.h
@@ -12,7 +12,7 @@ typedef void (*AfcReadCallback)(const void *reader_ptr, int64_t offset,
                                 int32_t *out_len);
 
 QImage generate_thumbnail_with_reader_ffi(const void *reader_ptr,
-                                          int32_t file_size,
+                                          int64_t file_size,
                                           int32_t requested_w,
                                           int32_t requested_h);
